requirements.txt
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
gspread==5.10.0
idna==3.4
numpy==1.25.1
oauthlib==3.2.2
pandas==2.0.3
PuLP==2.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
tzdata==2023.3
urllib3==1.26.16
